refactor(api/src/models): Renombrar archivos "ntf" a "nft" api working
